packages/image-modification-algorithms/pyproject.toml
[project]
name = "image-modification-algorithms"
version = "0.1.0"
description = "Reversible image modification algorithms"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.0",
    "numpy>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
